refactor: no-import component dep
