+++ src/components/Cards/PostCard/index.tsx
@@ -10,9 +10,9 @@ import {
   CardHeader, IconButton,
   Typography
 } from "@mui/material";
-import { Artist } from "../../constants/artist";
-import { Post } from "../../constants/post";
-import { useRouting } from "../../routes/useRouting";
+import { Artist } from "../../../constants/artist";
+import { Post } from "../../../constants/post";
+import { useRouting } from "../../../routes/useRouting";
 
 type Props = {
   post: Post,
